Clean part 1
